Fix PartsServiceTest test suite - Fix tests for the PartService by using AutoMapper dependency injection - Removes old test project Visual Studio for Mac failed to remove
