--- PartsTrader.ClientTools.MSTest/Repository/PartsRepositoryTest.cs
@@ -1,10 +1,0 @@
-﻿using System;
-namespace PartsTrader.ClientTools.MSTest.Repository
-{
-    public class PartsRepositoryTest
-    {
-        public PartsRepositoryTest()
-        {
-        }
-    }
-}
